Follow the standard file structure for CI with rake-compiler
